incred_urlbot/heroku/requirements.txt
pyTelegramBotAPI==3.7.6
PyGithub==1.54.1
validators==0.18.2
flask==1.1.2